minor simplification

+++ visualvm/libs.profiler/lib.profiler/src/org/graalvm/visualvm/lib/jfluid/heap/ObjectArrayDump.java
@@ -60,9 +60,9 @@ class ObjectArrayDump extends ArrayDump implements ObjectArrayInstance {
     //~ Methods ------------------------------------------------------------------------------------------------------------------
 
     public long getSize() {
-        int idSize = dumpClass.getHprofBuffer().getIDSize();
+        long idSize = dumpClass.getHprofBuffer().getIDSize();
 
-        return dumpClass.classDumpSegment.getMinimumInstanceSize() + HPROF_ARRAY_OVERHEAD + ((long)idSize * getLength());
+        return dumpClass.classDumpSegment.getMinimumInstanceSize() + HPROF_ARRAY_OVERHEAD + (idSize * getLength());
     }
 
     public List /*<Instance>*/ getValues() {
